set up react-native navigator
