+++ Test/components/welcome/Welcome.js
@@ -1,0 +1,45 @@
+import React, { Component } from 'react';
+import {
+  Platform,
+  StyleSheet,
+  Text,
+  Button,
+  View
+} from 'react-native';
+import  {StackNavigator} from 'react-navigation';
+
+export default class Welcome extends Component {
+    render() {
+      return (
+        <View style={styles.container}>
+          <Text style={styles.welcome}>
+            Welcome to welcome!
+          </Text>
+          <Button 
+          onPress={()=> {this.props.navigation.navigate('Login')}}
+          title="Sign Out"
+          >
+          </Button>
+        </View>
+      );
+    }
+  }
+  
+  const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      justifyContent: 'center',
+      alignItems: 'center',
+      backgroundColor: '#F5FCFF',
+    },
+    welcome: {
+      fontSize: 20,
+      textAlign: 'center',
+      margin: 10,
+    },
+    instructions: {
+      textAlign: 'center',
+      color: '#333333',
+      marginBottom: 5,
+    },
+  });
