+++ Test/package.json
@@ -9,7 +9,8 @@
   "dependencies": {
     "react": "16.3.0-alpha.1",
     "react-native": "0.54.1",
-    "react-native-google-signin": "^0.12.0"
+    "react-native-google-signin": "^0.12.0",
+    "react-navigation": "^1.5.2"
   },
   "devDependencies": {
     "babel-jest": "22.4.1",
